requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
bcrypt==3.2.0
blinker==1.4
boto3==1.21.18
botocore==1.24.18
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.7.4
fastapi==0.75.0
fastapi-mail==1.0.8
google-auth==2.6.6
h11==0.12.0
httpcore==0.14.7
httptools==0.2.0
httpx==0.22.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==0.10.0
loguru==0.6.0
MarkupSafe==2.1.0
mergedeep==1.3.4
motor==2.5.1
orjson==3.6.7
packaging==21.3
passlib==1.7.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pycparser==2.21
pydantic==1.8.2
pymongo==3.12.3
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
redis==4.2.2
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.17.1
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.2
wrapt==1.14.1
